fix: stock negativo y audit trail de inventario en ventas
- createSale registra InventoryMovement tipo egreso por cada producto vendido
- cancelSale usa LOCK.UPDATE + transaction en lectura de inventario y registra movimiento ingreso
- createTransfer agrega LOCK.UPDATE para evitar race condition entre transferencias concurrentes
- approveCreditNote: parseInt → parseFloat para no truncar decimales en devoluciones

+++ backend/controllers/creditNote.controller.js
@@ -576,7 +576,7 @@ exports.approveCreditNote = async (req, res) => {
 
         // Update inventory - ADD returned quantity
         await inventory.update({
-          quantity: parseInt(inventory.quantity) + totalUnits
+          quantity: parseFloat(inventory.quantity) + totalUnits
         }, { transaction });
 
         // Create inventory movement record
